requirements.txt
Django==3.2
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg==1.20.0
pre-commit==2.18.1
black==22.3.0
flake8==4.0.1